+++ plugins/dates/src/02-ranges/index.js
@@ -36,6 +36,7 @@ const parseRange = function (doc, context) {
     res = {
       start: unit,
       end: end,
+      unit: unit.unit,
     }
   }
   let combined = Object.assign({}, interval, res)